move spack.llnl.string to spack.util.string
