+++ lib/spack/spack/build_environment.py
@@ -84,7 +84,7 @@ from spack import traverse
 from spack.enums import Context
 from spack.error import InstallError, NoHeadersError, NoLibrariesError
 from spack.install_test import spack_install_test_log
-from spack.llnl.string import plural
+from spack.util.string import plural
 from spack.llnl.util.filesystem import join_path, symlink
 from spack.llnl.util.lang import dedupe, stable_partition
 from spack.llnl.util.tty.color import cescape, colorize
