+++ lib/spack/spack/filesystem_view.py
@@ -23,7 +23,7 @@ import spack.store
 import spack.util.spack_json as s_json
 import spack.util.spack_yaml as s_yaml
 from spack.error import SpackError
-from spack.llnl.string import comma_or
+from spack.util.string import comma_or
 from spack.llnl.util import tty
 from spack.llnl.util.filesystem import (
     mkdirp,
