+++ lib/spack/spack/llnl/util/lock.py
@@ -13,7 +13,7 @@ from typing import IO, Callable, Dict, Generator, Optional, Tuple, Type, Union
 
 from spack.llnl.util import lang, tty
 
-from ..string import plural
+from spack.util.string import plural
 
 if sys.platform != "win32":
     import fcntl
