+++ lib/spack/spack/stage.py
@@ -17,7 +17,7 @@ from typing import TYPE_CHECKING, Callable, Dict, Generator, Iterable, List, Opt
 import spack.caches
 import spack.config
 import spack.error
-import spack.llnl.string
+import spack.util.string
 import spack.llnl.util.lang
 import spack.llnl.util.tty as tty
 import spack.oci.image
@@ -1099,16 +1099,16 @@ def interactive_version_filter(
             header = []
             if len(orig_url_dict) > 0 and len(sorted_and_filtered) == len(orig_url_dict):
                 header.append(
-                    f"Selected {spack.llnl.string.plural(len(sorted_and_filtered), 'version')}"
+                    f"Selected {spack.util.string.plural(len(sorted_and_filtered), 'version')}"
                 )
             else:
                 header.append(
                     f"Selected {len(sorted_and_filtered)} of "
-                    f"{spack.llnl.string.plural(len(orig_url_dict), 'version')}"
+                    f"{spack.util.string.plural(len(orig_url_dict), 'version')}"
                 )
             if sorted_and_filtered and known_versions:
                 num_new = sum(1 for v in sorted_and_filtered if v not in known_versions)
-                header.append(f"{spack.llnl.string.plural(num_new, 'new version')}")
+                header.append(f"{spack.util.string.plural(num_new, 'new version')}")
             if has_filter:
                 header.append(colorize(f"Filtered by {VERSION_COLOR}@@{version_filter}@."))
