+++ lib/spack/spack/util/string.py
